fixed reference error to species

+++ backend/WhaleSpotting/Models/Request/ModifyPostRequest.cs
@@ -4,8 +4,8 @@ public class ModifyPostRequest
 {
     public int id { get; set; }
     public DateTime date { get; set; }
-    public int lat { get; set; }
-    public int lon { get; set; }
+    public double lat { get; set; }
+    public double lon { get; set; }
     public int species { get; set; }
     public string description { get; set; }
     public string imageUrl { get; set; }
